refactor: move variant folder in correct location
